Shooting projectiles, crosshair, and additional controls
The player can now shoot basic projectiles from their plane, which can pop balloons objects on collision. Collision detection for the projectiles and balloons is (after a lot of debugging) very consistent and not dependent on distance or framerate. The plane has a crosshair to show where the projectiles are headed. The player can also hold shift to lower their mouse sensitivity, allowing for precise aiming. Finally, the player can hold ctrl to turn off the engine, but right now the physics don't work the way I'd like them to (the plane falls in a strange way, and does not allow the player to glide).