Tilt controls
The user can now hold A or D to tilt their plane left or right (roll rotation). This is a little nausea inducing though, so I may revisit this mechanic.

Leaving out a change to the "Library/ArtifactDB" file because it's too big to push to GitHub. Hope that won't cause problems! 🤞