Balloon Packages and Package Spawning
A large white balloon will now spawn in one of any number of predetermined locations chosen at random 5 seconds after the game starts. When the player collides with the balloon package, it is disabled and the player is rewarded with 10 more balloons. After the balloon package is collected, a new one will spawn 10 seconds later.