Crash detection and player health refactoring
The propeller now has a hitbox that when collided with, subtracts 10 balloons from the player's health.

BalloonHitbox has been deprecated (although still used for testing) and split into 3 different classes: Hitbox, PlayerHealth, and BalloonResize. They talk to each other using events. This allows for better following of the containment model and the single responsibility principle.

Finally, I also made the plane's propeller switch to the unblurred model when the air break is engaged (the engine is off and plane is gliding)